backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1

# 🔽 IMPORTANT FIXES
numpy==1.26.4
pandas==2.1.4
openpyxl==3.1.2
